+++ changelog.md
@@ -5,6 +5,16 @@ All changes of this project must be added in this file.
 The format for based in [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the project is follow the versioning according to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 16/09/2025
+
+### Added
+
+- Added support to multiple tests in one file
+- Added after command to run after test
+- Added reports to show in console and junit format
+- Added a complex test with multiple asserts
+- Added flag --log-level to set log level (debug, info, warn, error)
+
 ## [0.1.0] - 31/08/2024
 
 ### Added